+++ superduperdb/backends/base/artifacts.py
@@ -101,9 +101,13 @@ class ArtifactStore(ABC):
                   and optional fields
                   {'file_id', 'uri'}
         """
-        if r.get('leaf_type') != 'file' and 'file_id' in r and r.get('bytes') is None:
+        if (
+            r.get('artifact_type') != 'file'
+            and 'file_id' in r
+            and r.get('bytes') is None
+        ):
             return
-        if r.get('leaf_type') == 'file':
+        if r.get('artifact_type') == 'file':
             assert 'file_id' in r, 'file_id is missing!'
             file_id = self._save_file(r['uri'], r['file_id'])
         else:
@@ -171,6 +175,7 @@ class ArtifactStore(ABC):
                 'artifact',
                 'file',
                 'lazy_artifact',
+                'lazy_file',
             }:
                 self.save_artifact(r['_content'])
             else: